% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_attractiveness.R
\name{calculate_attractiveness}
\alias{calculate_attractiveness}
\title{Calculate Attractiveness Score}
\usage{
calculate_attractiveness(
  data,
  weights = list(COL = 0.6, Visa = 0.4, Tax = 0.2, Dist = 0.2, Talent = 0.6)
)
}
\arguments{
\item{data}{A data frame with required columns.}

\item{weights}{A list of weights for each metric.}
}
\value{
A data frame with Attractiveness_Score added.
}
\description{
Calculate Attractiveness Score
}
