% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_regions.R
\name{compare_regions}
\alias{compare_regions}
\title{Compare Two Regions}
\usage{
compare_regions(data, region1, region2)
}
\arguments{
\item{data}{A data frame with Attractiveness_Score calculated.}

\item{region1}{The first region to compare.}

\item{region2}{The second region to compare.}
}
\value{
A comparison message.
}
\description{
This function compares the attractiveness scores of two regions.
}
