% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_regions.R
\name{rank_regions}
\alias{rank_regions}
\title{Rank Regions by Attractiveness Score}
\usage{
rank_regions(calculated_data, top_n = NULL)
}
\arguments{
\item{calculated_data}{A data frame with Attractiveness_Score.}

\item{top_n}{Number of top regions to return.}
}
\value{
A ranked data frame.
}
\description{
Rank Regions by Attractiveness Score
}
