% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_weights.R
\name{tune_weights}
\alias{tune_weights}
\title{Tune Weights for Attractiveness Score}
\usage{
tune_weights(
  data,
  weights = list(COL = 0.6, Visa = 0.4, Tax = 0.2, Dist = 0.2, Talent = 0.6)
)
}
\arguments{
\item{data}{A data frame with required columns for metrics.}

\item{weights}{A list of initial weights.}
}
\value{
A data frame with recalculated Attractiveness_Score.
}
\description{
Allows users to adjust weights and recalculate attractiveness scores.
}
